+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "sweepai"
-version = "0.2.0"
+version = "0.2.1"
 description = "Sweep software chores"
 authors = ["Kevin Lu", "William Zeng", "Luke Jagg"]
 packages = [{ include = "sweepai" }]
